+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## 1.1.0 (16th June, 2021)
+
+New:
+
+- Added support for setting an owner against an imported document
+
+Fixed:
+
+- Issue where tags containing backslash characters could not be found
+
 ## 1.0.0 (17th January, 2020)
 
 Initial Release